feature/SMP-133 created PauseWidget class which provides all pause game utilities
