+++ core/src/com/group/gra/screens/GameScreen.java
@@ -1,6 +1,6 @@
 package com.group.gra.screens;
 
-import com.badlogic.gdx.Game;
+
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Preferences;
 import com.badlogic.gdx.Screen;
@@ -16,7 +16,6 @@ import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.actions.Actions;
 import com.badlogic.gdx.scenes.scene2d.actions.SequenceAction;
 import com.badlogic.gdx.scenes.scene2d.ui.*;
-import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
 import com.badlogic.gdx.utils.Array;
 import com.badlogic.gdx.utils.viewport.FitViewport;
 import com.group.gra.DragAndDropNew;
@@ -45,10 +44,7 @@ public class GameScreen implements Screen {
     private Integer counter;
     private Label label;
     private Skin skin;
-    private Window windowPause;
-    private boolean pause;
-    final Button buttonPause,buttonSound;
-    final TextButton buttonResume,buttonQuit;
+    private PauseWidget pauseWidget;
 
     public GameScreen(SpriteBatch sb) {
         this.sb = sb;
@@ -70,33 +66,18 @@ public class GameScreen implements Screen {
 
         TextureAtlas atlas = new TextureAtlas("ui/design.atlas");
         skin = new Skin(Gdx.files.internal("ui/design.json"), atlas);
-
-        buttonPause = new Button(skin.get("pause",Button.ButtonStyle.class));
-        buttonPause.setSize(50, 50);
-        addButtonPauseListener(buttonPause);
-        stage.addActor(buttonPause);
-
-        buttonResume = new TextButton("Resume",skin);
-        addButtonResume(buttonResume);
-        buttonQuit = new TextButton("Quit",skin);
-        addButtonQuit(buttonQuit);
-
-
         Preferences prefs = Gdx.app.getPreferences(SETTINGS_FILE);
-        buttonSound = createButtonSound(prefs);
-        buttonSound.setBounds(stage.getWidth()-50,stage.getHeight()-50,50,50);
-        addButtonSoundListener(buttonSound);
 
-        windowPause = new Window("", skin);
-        windowPause.setSize(stage.getWidth(),stage.getHeight());
-        windowPause.setMovable(false);
-        Table table = new Table(skin);
-        windowPause.add(table);
+        backgroundMusic = Gdx.audio.newMusic(Gdx.files.internal("music/gameScreenMusic.mp3"));
+        backgroundMusic.setLooping(true);
+        backgroundMusic.setVolume(0.1f);
 
-        table.add(buttonResume).row();
-        table.add(buttonQuit).row();
-        table.add(buttonSound).row();
+        if(prefs.getBoolean(SOUND_ON)) {
+            backgroundMusic.play();
+        }
 
+        pauseWidget = new PauseWidget(backgroundMusic,stage,sb);
+        pauseWidget.createButtonPause(0, 0, 50, 50);
 
         TrashGenerator generator = new TrashGenerator();
         Array<Trash> trashArray = generator.generateTrashArray(10);
@@ -114,15 +95,6 @@ public class GameScreen implements Screen {
             trash.getImage().addAction(sequence);
             DELAY_TIME += 2;
         }
-
-        backgroundMusic = Gdx.audio.newMusic(Gdx.files.internal("music/gameScreenMusic.mp3"));
-        backgroundMusic.setLooping(true);
-        backgroundMusic.setVolume(0.1f);
-
-        if(prefs.getBoolean(SOUND_ON)) {
-            backgroundMusic.play();
-        }
-
     }
 
     private Label createCounterLabel() {
@@ -224,7 +196,7 @@ public class GameScreen implements Screen {
         spriteBackground.draw(sb);
         maciag.draw(sb);
         sb.end();
-        if(pause) {
+        if(pauseWidget.isGamePaused()) {
             delta = 0;
         }
         stage.draw();
@@ -261,59 +233,4 @@ public class GameScreen implements Screen {
         backgroundMusic.dispose();
     }
 
-
-    private void addButtonPauseListener(Button buttonPause) {
-        buttonPause.addListener(new ClickListener() {
-            @Override
-            public void clicked(InputEvent event, float x, float y) {
-                pause=true;
-                stage.addActor(windowPause);
-            }
-        });
-    }
-
-    private void addButtonResume(final TextButton buttonResume) {
-        buttonResume.addListener(new ClickListener() {
-            @Override
-            public void clicked(InputEvent event, float x, float y) {
-                pause=false;
-                windowPause.addAction(Actions.removeActor());
-            }
-        });
-    }
-
-    private void addButtonQuit(final TextButton buttonQuit) {
-        buttonQuit.addListener(new ClickListener() {
-            @Override
-            public void clicked(InputEvent event, float x, float y) {
-                ((Game) Gdx.app.getApplicationListener()).setScreen(new MenuScreen(sb));
-            }
-        });
-    }
-
-    private void addButtonSoundListener(final Button buttonSound) {
-        buttonSound.addListener(new ClickListener() {
-            @Override
-            public void clicked(InputEvent event, float x, float y) {
-                Preferences prefs = Gdx.app.getPreferences(SETTINGS_FILE);
-                if (prefs.getBoolean(SOUND_ON,true)) {
-                    prefs.putBoolean(SOUND_ON, false).flush();
-                    buttonSound.setStyle(skin.get("sound_off",Button.ButtonStyle.class));
-                    backgroundMusic.stop();
-                }
-                else {
-                    prefs.putBoolean(SOUND_ON, true).flush();
-                    buttonSound.setStyle(skin.get("sound_on",Button.ButtonStyle.class));
-                    backgroundMusic.play();
-                }
-            }
-        });
-    }
-    private Button createButtonSound(Preferences prefs) {
-        if (prefs.getBoolean(SOUND_ON,true)) {
-            return new Button(skin.get("sound_on",Button.ButtonStyle.class));
-        } else {
-            return new Button(skin.get("sound_off",Button.ButtonStyle.class));
-        }
-    }
 }
